requirements.txt
numpy==1.19.2
pandas
matplotlib
scipy==1.4.1
h5py<3.0.0
keras==2.3.1
PyYaml
RootInteractive
tensorflow>=2.4.0
pydot
pylint
notebook
torch
xgboost
